+++ src/index.ts
@@ -5,11 +5,11 @@ export * from "./currency";
 export * from "./invoice";
 export * from "./membership";
 export * from "./product";
+export * from "./purchase-order";
 export * from "./sales-order";
 export * from "./shipping-policy";
 export * from "./store";
 export * from "./subscription";
-export * from "./purchaseorder";
 export * from "./supplier";
 export * from "./tax";
 export * from "./webhook";
